[SIG-24903] Prototype for Snowflake SubmitSync (#81)
Adds a SubmitSync method to the Snowflake driver that directly calls Snowflake's POST /queries/v1/query-request API endpoint synchronously, waiting for up to 45 seconds (the current, fixed timeout) but does not enter the "ping pong" phase of fetching query results for long-running queries. Instead, the caller is responsible for using the query ID to wait for and fetch query results.

This prototype mainly exists right now so that we can test the relevant functionality and compare performance with the fully asynchronous mode of execution.